+++ CHANGELOG.md
@@ -1,4 +1,13 @@
 # Changelog
+## [v5.4.3] - 2026-06-17
+## Fixed
+- Fix camera issue in naval battle when piloting ship. Caused by elevated camera in v5.4.2.
+
+### Improved
+- Improve behavior of elevated camera.
+
+- Elevated height is saved in configuration.
+
 ## [v5.4.2] - 2026-06-16
 ## Added
 - When giving orders, camera will be elevated. Corresponding option is added.